test geolocation e2e with a test mmdb
