+++ config/.env.test
@@ -13,3 +13,4 @@ SELFHOST=false
 SITE_LIMIT=3
 HCAPTCHA_SITEKEY=test
 HCAPTCHA_SECRET=scottiger
+IP_GEOLOCATION_DB=test/priv/GeoLite2-City-Test.mmdb
