+++ test/plausible/ingestion/event_test.exs
@@ -15,7 +15,7 @@ defmodule Plausible.Ingestion.EventTest do
   end
 
   @valid_request %Plausible.Ingestion.Request{
-    remote_ip: "2.2.2.2",
+    remote_ip: "2.125.160.216",
     user_agent:
       "Mozilla/5.0 (iPad; U; CPU OS 3_2_1 like Mac OS X; en-us) AppleWebKit/531.21.10 (KHTML, like Gecko) Mobile/7B405",
     event_name: "pageview",
@@ -46,8 +46,8 @@ defmodule Plausible.Ingestion.EventTest do
              domain: "plausible-ingestion-event-basic.test",
              browser: "Safari",
              browser_version: "",
-             city_geoname_id: 2_988_507,
-             country_code: "FR",
+             city_geoname_id: 2_655_045,
+             country_code: "GB",
              hostname: "skywalker.test",
              "meta.key": [],
              "meta.value": [],
@@ -58,8 +58,8 @@ defmodule Plausible.Ingestion.EventTest do
              referrer: "m.facebook.test",
              referrer_source: "utm_source",
              screen_size: "Desktop",
-             subdivision1_code: "FR-IDF",
-             subdivision2_code: "FR-75",
+             subdivision1_code: "GB-ENG",
+             subdivision2_code: "GB-WBK",
              transferred_from: "",
              utm_campaign: "utm_campaign",
              utm_content: "utm_content",
