+++ test/plausible_web/controllers/api/external_controller_test.exs
@@ -630,7 +630,7 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       assert event.referrer == ""
     end
 
-    # Fake data is set up in config/test.exs
+    # Fake geo is loaded from test/priv/GeoLite2-City-Test.mmdb
     test "looks up location data from the ip address", %{conn: conn} do
       params = %{
         name: "pageview",
@@ -639,15 +639,15 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       }
 
       conn
-      |> put_req_header("x-forwarded-for", "2.2.2.2")
+      |> put_req_header("x-forwarded-for", "2.125.160.216")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-20.com")
 
-      assert pageview.country_code == "FR"
-      assert pageview.subdivision1_code == "FR-IDF"
-      assert pageview.subdivision2_code == "FR-75"
-      assert pageview.city_geoname_id == 2_988_507
+      assert pageview.country_code == "GB"
+      assert pageview.subdivision1_code == "GB-ENG"
+      assert pageview.subdivision2_code == "GB-WBK"
+      assert pageview.city_geoname_id == 2_655_045
     end
 
     test "ignores unknown country code ZZ", %{conn: conn} do
@@ -674,7 +674,7 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       }
 
       conn
-      |> put_req_header("x-forwarded-for", "1.1.1.1:123")
+      |> put_req_header("x-forwarded-for", "216.160.83.56:123")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-x-forwarded-for-port.com")
@@ -690,12 +690,12 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       }
 
       conn
-      |> put_req_header("x-forwarded-for", "1:1:1:1:1:1:1:1")
+      |> put_req_header("x-forwarded-for", "2001:218:1:1:1:1:1:1")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-x-forwarded-for-ipv6.com")
 
-      assert pageview.country_code == "US"
+      assert pageview.country_code == "JP"
     end
 
     test "works with ipv6 with a port number in x-forwarded-for", %{conn: conn} do
@@ -706,12 +706,12 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       }
 
       conn
-      |> put_req_header("x-forwarded-for", "[1:1:1:1:1:1:1:1]:123")
+      |> put_req_header("x-forwarded-for", "[2001:218:1:1:1:1:1:1]:123")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-x-forwarded-for-ipv6-port.com")
 
-      assert pageview.country_code == "US"
+      assert pageview.country_code == "JP"
     end
 
     test "uses cloudflare's special header for client IP address if present", %{conn: conn} do
@@ -723,7 +723,7 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
 
       conn
       |> put_req_header("x-forwarded-for", "0.0.0.0")
-      |> put_req_header("cf-connecting-ip", "1.1.1.1")
+      |> put_req_header("cf-connecting-ip", "216.160.83.56")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-cloudflare.com")
@@ -740,7 +740,7 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
 
       conn
       |> put_req_header("x-forwarded-for", "0.0.0.0")
-      |> put_req_header("b-forwarded-for", "1.1.1.1,9.9.9.9")
+      |> put_req_header("b-forwarded-for", "216.160.83.56,9.9.9.9")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-bunny.com")
@@ -758,7 +758,7 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       }
 
       conn
-      |> put_req_header("forwarded", "by=0.0.0.0;for=1.1.1.1;host=somehost.com;proto=https")
+      |> put_req_header("forwarded", "by=0.0.0.0;for=216.160.83.56;host=somehost.com;proto=https")
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-forwarded.com")
@@ -776,13 +776,13 @@ defmodule PlausibleWeb.Api.ExternalControllerTest do
       conn
       |> put_req_header(
         "forwarded",
-        "by=0.0.0.0;for=\"[1:1:1:1:1:1:1:1]\",for=0.0.0.0;host=somehost.com;proto=https"
+        "by=0.0.0.0;for=\"[2001:218:1:1:1:1:1:1]\",for=0.0.0.0;host=somehost.com;proto=https"
       )
       |> post("/api/event", params)
 
       pageview = get_event("external-controller-test-forwarded-ipv6.com")
 
-      assert pageview.country_code == "US"
+      assert pageview.country_code == "JP"
     end
 
     test "URL is decoded", %{conn: conn} do
