+++ test/priv/README.md
@@ -1,0 +1,1 @@
+`GeoLite2-City-Test.mmdb` is downloaded from https://github.com/maxmind/MaxMind-DB
